Began designing simulation
The script begins by creating the cards in a deck. The deck chooses randomly how many lands (between 1 and 59, since 0 and 60 will leave no possibility of ever winning) will be in the deck, then adds cards with random costs until the deck is full. It shuffles the 'deck' and removes the top 7 cards to be part of the hand. I've begun to design a standard turn procedure, which involves playing a land if possible, then playing a creature or creatures whose total cost is equal to the number of lands on the battlefield.